--- src/main/kotlin/kr/kro/dearmoment/like/adapter/input/web/dto/LikeResponse.kt
@@ -1,5 +1,0 @@
-package kr.kro.dearmoment.like.adapter.input.web.dto
-
-data class LikeResponse(
-    val likeId: Long,
-)
