+++ src/main/kotlin/kr/kro/dearmoment/like/application/dto/LikeResponse.kt
@@ -1,0 +1,5 @@
+package kr.kro.dearmoment.like.application.dto
+
+data class LikeResponse(
+    val likeId: Long,
+)
