+++ src/main/kotlin/kr/kro/dearmoment/like/application/port/input/LikeUseCase.kt
@@ -1,8 +1,8 @@
 package kr.kro.dearmoment.like.application.port.input
 
-import kr.kro.dearmoment.like.adapter.input.web.dto.GetLikeResponse
-import kr.kro.dearmoment.like.adapter.input.web.dto.LikeResponse
 import kr.kro.dearmoment.like.application.command.SaveLikeCommand
+import kr.kro.dearmoment.like.application.dto.GetLikeResponse
+import kr.kro.dearmoment.like.application.dto.LikeResponse
 import kr.kro.dearmoment.like.application.query.ExistLikeQuery
 import kr.kro.dearmoment.like.application.query.GetLikesQuery
